Update patch set 1
Patch Set 1: Fails; I would prefer that you didn't submit this

Unfortunately the code doesn't merge cleanly onto the latest master HEAD. If it's still relevant, please resolve the conflicts and upload a replacement.

I was wondering, would it be possible to map ppp log levels to android log levels in the logit() function instead of spreading the mapping around throughout the rest of the code? The drawback would be that it'd be a runtime mapping instead of compile-time.

Patch-set: 1
Reviewer: Jean-Baptiste Queru <1000660@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Verified=-1