Update patch set 2
Patch Set 2:

We did not support this from the beginning because of some security concerns. However, we did receive the requests and understand the needs from advanced and enterprise users.

The java part looks fine, and Hung-ying also gave me a green light for that. For the ppp part, things change a lot since donut, so I might do it for you. Maybe it should be done in netd, or maybe not. I also need to consult the security council, so please stay tuned.

Patch-set: 2