Update patch set 3
Patch Set 3:

(2 comments)

The remaining diff to upstream (for the bits that we compile) is:

https://paste.googleplex.com/5069590618963968

I had to readd the DesEncrypt change because there are two versions of that function and upstream have fixed one, but not the one that we currently use.

I also added __ANDROID__ tags around a local change that I had missed (in utils.c).

Patch-set: 3