pppd: Add rules for building the pppol2tp plugin
Build plugin as shared library. It uses some symbols from pppd, so let's
allow building the plugin with undefined symbols, as they will be
resolved at runtime, when plugin is loaded (on dlopen() call).

Plugin will be installed to next path:
  - on 32-bit systems: /system/lib/
  - on 64-bit systems: /system/lib64/
  - on Multilib build: to both paths above

We install it in /system/$LIB/ rather than /system/$LIB/pppd/$VERSION/,
because we previously modified pppd to look in /system/$LIB for plugins,
as it is the only allowed path for dlopen().

While at it, extract flags that common for pppd and plugins to separate
variable, to avoid code duplication.

Also, add pppol2tp as pppd dependency, so it will be built and installed
along with pppd. Let's use LOCAL_REQUIRED_MODULE instead of
LOCAL_SHARED_LIBRARIES, because latter would lead to linking pppd binary
against pppol2tp.so, which we don't want, as pppol2tp.so is just a plugin
and should be loaded via dlopen() only when it's needed.

Change-Id: I4b51a0f2bf99550c0fe126b6e56e08127dde812b
Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>