pppd: Add rules for building the pppol2tp-android plugin
Build plugin as shared library. It uses some symbols from pppd, so let's
allow building the plugin with undefined symbols, as they will be
resolved at runtime, when plugin is loaded (on dlopen() call).

Plugin will be installed to next path:
  - on 32-bit systems: /system/lib/
  - on 64-bit systems: /system/lib64/
  - on Multilib build: to both paths above

We install it in /system/$LIB/ rather than /system/$LIB/pppd/$VERSION/,
because we previously modified pppd to look in /system/$LIB for plugins,
as it is the only allowed path for dlopen().

While at it, extract flags that common for pppd and plugins to separate
variable, to avoid code duplication.

Also, add pppol2tp-android as pppd dependency, so it will be built and
installed along with pppd. Let's use LOCAL_REQUIRED_MODULE instead of
LOCAL_SHARED_LIBRARIES, because latter would lead to linking pppd binary
against pppol2tp-android.so, which we don't want, as pppol2tp-android.so
is just a plugin and should be loaded via dlopen() only when it's needed.

Enable USING_UAPI for pppol2tp-android build. When trying to build
pppol2tp-android.so as is (in current pppd codebase) we will see build
errors like this one:

    bionic/libc/kernel/uapi/linux/ppp_defs.h:59:6:
        error: redefinition of 'NPmode'
    external/ppp/pppd/include/net/ppp_defs.h:142:6:
        note: previous definition is here

This is basically because USING_API option is not defined when building
pppol2tp-android.c, though it should be. Enabling that option fixes the
build. Detailed explanation is below.

Unlike the upstream pppd project, we don't keep include/linux/* headers
inside of pppd codebase. Instead, we define USING_UAPI when such headers
(user-space API [1]) are used, providing missing functionality in
ppp_defs.h file. From pppol2tp-android.c file it's obvious that UAPI is
used there. So let's define USING_UAPI for the whole pppol2tp-android
project, instead of trying to check if we need to define it in compile
time.

-Wno-pointer-sign is added to avoid next error shown by presubmit
(Treehugger Robot), because 32-bit ABI is broken:

        error: passing 'int *' to parameter of type 'socklen_t *'
        (aka 'unsigned int *') converts between pointers to integer
        types with different sign [-Werror,-Wpointer-sign]

[1] https://lwn.net/Articles/507794/

Change-Id: I74fe232366fa496662323c669c02dffb5c8ecd53
Signed-off-by: Sam Protsenko <semen.protsenko@linaro.org>