Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Yeah, those patches seem like they make more sense to me.
> 
> Have you talked to upstream? Do we know whether they'd be happy with those new patches, or would prefer to have a separate #if __ANDROID__? (Since the goal is to unify AOSP and Android. I don't know whether they're a "whatever works" kind of project or likely to have strong opinions.)

I'll do that. But that discussion can take some time (I see that last commit was done a year ago to ppp project). Is that possible for us to try and merge that patch series and continue our discussion with ppp project after that? As I see it, this chunk of work (use upstream L2TP) is functional. Would be nice to do upstream-related work as a step 2, after merging this series. Also, once it's merged, I think I can start doing the same work for PPTP. Is that all sound reasonable?

Patch-set: 2
Reviewer: Sam Protsenko <1131154@85c56323-6fa9-3386-8a01-6480fb634889>